troubleshooting data input
We noticed nulls coming in on the start_time fields. We chose to look into this. And realized that new csv files have a different column name for start time than old csv files and that is the cause for the issue. I have put this code here in case i want to ever dig further into any other troubleshooting items on the data pipeline